requirements.txt
pyyaml>=5.4.1
yfinance>=0.2.0
pandas>=1.3.0

